chore: split server-adapter package
