+++ packages/server-middleware-http/src/index.ts
@@ -1,9 +1,9 @@
 /*
- * Copyright (c) 2022.
+ * Copyright (c) 2022-2024.
  * Author Peter Placzek (tada5hi)
  * For the full copyright and license information,
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './middleware';
-export * from './verifier';
+export * from './module';
+export * from './type';
