+++ packages/server-middleware-http/src/type.ts
@@ -1,18 +1,18 @@
 /*
- * Copyright (c) 2022.
+ * Copyright (c) 2022-2024.
  * Author Peter Placzek (tada5hi)
  * For the full copyright and license information,
  * view the LICENSE file that was distributed with this source code.
  */
 
 import type { IncomingMessage, ServerResponse } from 'node:http';
-import type { TokenVerificationData, TokenVerifier, TokenVerifierOptions } from '../../verifier';
+import type { TokenVerificationData, TokenVerifier, TokenVerifierOptions } from '@authup/server-middleware-kit';
 
-export type HTTPMiddlewareOptions = {
+export type MiddlewareOptions = {
     tokenByCookie?: (req: IncomingMessage, name?: string) => string | undefined,
     tokenVerifier: TokenVerifier | TokenVerifierOptions,
     tokenVerifierHandler: (req: IncomingMessage, data: TokenVerificationData) => void
 };
 
-export type HTTPNext = (err?: Error) => void;
-export type HTTPMiddleware = (req: IncomingMessage, res: ServerResponse, next: HTTPNext) => any;
+export type Next = (err?: Error) => void;
+export type Middleware = (req: IncomingMessage, res: ServerResponse, next: Next) => any;
