+++ packages/server-middleware-kit/package.json
@@ -1,0 +1,71 @@
+{
+    "name": "@authup/server-middleware-kit",
+    "version": "1.0.0-beta.9",
+    "description": "A package containing utilities for middleware solutions.",
+    "exports": {
+        "./package.json": "./package.json",
+        ".": {
+            "types": "./dist/index.d.ts",
+            "import": "./dist/index.mjs",
+            "require": "./dist/index.cjs"
+        }
+    },
+    "main": "./dist/index.cjs",
+    "module": "./dist/index.mjs",
+    "types": "./dist/index.d.ts",
+    "files": [
+        "dist/"
+    ],
+    "scripts": {
+        "build:types": "tsc --emitDeclarationOnly -p tsconfig.build.json",
+        "build:js": "rollup -c",
+        "build": "rimraf ./dist && cross-env NODE_ENV=production npm run build:js && npm run build:types",
+        "test": "cross-env NODE_TLS_REJECT_UNAUTHORIZED=0 NODE_ENV=test jest --config ./test/jest.config.js",
+        "test:coverage": "npm run test -- --coverage"
+    },
+    "keywords": [
+        "auth",
+        "module",
+        "typescript",
+        "middleware",
+        "http",
+        "express",
+        "socket",
+        "socket.io",
+        "websocket",
+        "adapter",
+        "oauth2"
+    ],
+    "author": {
+        "name": "Peter Placzek",
+        "email": "contact@tada5hi.net",
+        "url": "https://github.com/tada5hi"
+    },
+    "license": "Apache-2.0",
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/authup/authup.git",
+        "directory": "server/packages/adapter"
+    },
+    "bugs": {
+        "url": "https://github.com/authup/authup/issues"
+    },
+    "homepage": "https://github.com/authup/authup#readme",
+    "dependencies": {
+        "jose": "^5.2.4",
+        "redis-extension": "^1.3.0"
+    },
+    "devDependencies": {
+        "@authup/core-kit": "^1.0.0-beta.9",
+        "@authup/server-kit": "^1.0.0-beta.9",
+        "@hapic/oauth2": "^2.4.2"
+    },
+    "peerDependencies": {
+        "@authup/core-kit": "^1.0.0-beta.9",
+        "@authup/server-kit": "^1.0.0-beta.9"
+    },
+    "publishConfig": {
+        "access": "public"
+    },
+    "gitHead": "ed48328bdd01d844915c26c6fcfb7fdcc1f33396"
+}
