+++ packages/server-middleware-kit/src/index.ts
@@ -5,5 +5,4 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './module';
-export * from './type';
+export * from './verifier';
