+++ packages/server-middleware-socket/README.md
@@ -1,0 +1,28 @@
+# @authup/server-adapter 🌉
+
+[![npm version](https://badge.fury.io/js/@authup%2Fserver-adapter.svg)](https://badge.fury.io/js/@authup%2Fserver-adapter)
+[![main](https://github.com/authup/authup/actions/workflows/main.yml/badge.svg)](https://github.com/authup/authup/actions/workflows/main.yml)
+[![codecov](https://codecov.io/gh/authup/authup/branch/master/graph/badge.svg?token=FHE347R1NW)](https://codecov.io/gh/authup/authup)
+[![Known Vulnerabilities](https://snyk.io/test/github/authup/authup/badge.svg)](https://snyk.io/test/github/authup/authup)
+
+The main propose of this package, is to provide middlewares for:
+
+- http &
+- (web-) socket
+
+based services. These middlewares provide a way to validate and verify the request and inject information about the corresponding token owner.
+
+**Table of Contents**
+
+- [Documentation](#documentation)
+- [License](#license)
+
+## Documentation
+
+To read the docs, visit [https://authup.org/packages/server-adapter](https://authup.org/packages/server-adapter)
+
+## License
+
+Made with 💚
+
+Published under [Apache 2.0 License](./LICENSE).
