+++ packages/server-middleware-socket/package.json
@@ -1,0 +1,55 @@
+{
+    "name": "@authup/server-middleware-socket",
+    "version": "1.0.0-beta.9",
+    "description": "An adapter package to integrate into other services.",
+    "exports": {
+        "./package.json": "./package.json",
+        ".": {
+            "types": "./dist/index.d.ts",
+            "import": "./dist/index.mjs",
+            "require": "./dist/index.cjs"
+        }
+    },
+    "main": "./dist/index.cjs",
+    "module": "./dist/index.mjs",
+    "types": "./dist/index.d.ts",
+    "files": [
+        "dist/"
+    ],
+    "scripts": {
+        "build:types": "tsc --emitDeclarationOnly -p tsconfig.build.json",
+        "build:js": "rollup -c",
+        "build": "rimraf ./dist && cross-env NODE_ENV=production npm run build:js && npm run build:types"
+    },
+    "keywords": [],
+    "author": {
+        "name": "Peter Placzek",
+        "email": "contact@tada5hi.net",
+        "url": "https://github.com/tada5hi"
+    },
+    "license": "Apache-2.0",
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/authup/authup.git",
+        "directory": "packages/server-middleware-http"
+    },
+    "bugs": {
+        "url": "https://github.com/authup/authup/issues"
+    },
+    "homepage": "https://github.com/authup/authup#readme",
+    "dependencies": {
+        "@authup/server-middleware-kit": "^1.0.0-beta.9"
+    },
+    "devDependencies": {
+        "@authup/core-kit": "^1.0.0-beta.9",
+        "@authup/server-kit": "^1.0.0-beta.9"
+    },
+    "peerDependencies": {
+        "@authup/core-kit": "^1.0.0-beta.9",
+        "@authup/server-kit": "^1.0.0-beta.9"
+    },
+    "publishConfig": {
+        "access": "public"
+    },
+    "gitHead": "ed48328bdd01d844915c26c6fcfb7fdcc1f33396"
+}
