+++ packages/server-middleware-socket/src/type.ts
@@ -1,18 +1,18 @@
 /*
- * Copyright (c) 2022.
+ * Copyright (c) 2022-2024.
  * Author Peter Placzek (tada5hi)
  * For the full copyright and license information,
  * view the LICENSE file that was distributed with this source code.
  */
 
-import type { TokenVerificationData, TokenVerifier, TokenVerifierOptions } from '../../verifier';
+import type { TokenVerificationData, TokenVerifier, TokenVerifierOptions } from '@authup/server-middleware-kit';
 
-export type SocketMiddlewareOptions = {
+export type MiddlewareOptions = {
     tokenVerifier: TokenVerifierOptions | TokenVerifier,
     tokenVerifierHandler: (socket: Socket, data: TokenVerificationData) => void
 };
 
-export type SocketNextFunction = (err?: Error) => void;
+export type Next = (err?: Error) => void;
 
 export type Handshake = {
     auth: {
